refactor(store): remove legacy store (#7562)
Closes: BS-840
