+++ packages/blocks/src/__tests__/database/database.unit.spec.ts
@@ -1,6 +1,6 @@
 import type { BlockModel, Doc } from '@blocksuite/store';
 
-import { DocCollection, Generator, Schema } from '@blocksuite/store';
+import { DocCollection, IdGeneratorType, Schema } from '@blocksuite/store';
 import { beforeEach, describe, expect, test } from 'vitest';
 
 import type { DatabaseBlockModel } from '../../database-block/database-model.js';
@@ -25,7 +25,7 @@ const AffineSchemas = [
 ];
 
 function createTestOptions() {
-  const idGenerator = Generator.AutoIncrement;
+  const idGenerator = IdGeneratorType.AutoIncrement;
   const schema = new Schema();
   schema.register(AffineSchemas);
   return { id: 'test-collection', idGenerator, schema };
