+++ packages/framework/store/src/__tests__/schema.unit.spec.ts
@@ -7,7 +7,7 @@ import { describe, expect, it, vi } from 'vitest';
 import { type BlockModel, defineBlockSchema } from '../schema/base.js';
 import { SchemaValidateError } from '../schema/error.js';
 import { Schema } from '../schema/index.js';
-import { DocCollection, Generator } from '../store/index.js';
+import { DocCollection, IdGeneratorType } from '../store/index.js';
 import {
   DividerBlockSchema,
   ListBlockSchema,
@@ -17,7 +17,7 @@ import {
 } from './test-schema.js';
 
 function createTestOptions() {
-  const idGenerator = Generator.AutoIncrement;
+  const idGenerator = IdGeneratorType.AutoIncrement;
   const schema = new Schema();
   schema.register(BlockSchemas);
   return { id: 'test-collection', idGenerator, schema };
