+++ packages/framework/store/src/__tests__/transformer.unit.spec.ts
@@ -9,7 +9,7 @@ import {
   type SchemaToModel,
   defineBlockSchema,
 } from '../schema/index.js';
-import { DocCollection, Generator } from '../store/index.js';
+import { DocCollection, IdGeneratorType } from '../store/index.js';
 import { AssetsManager, BaseBlockTransformer } from '../transformer/index.js';
 
 const docSchema = defineBlockSchema({
@@ -44,7 +44,7 @@ const docSchema = defineBlockSchema({
 type RootBlockModel = SchemaToModel<typeof docSchema>;
 
 function createTestOptions() {
-  const idGenerator = Generator.AutoIncrement;
+  const idGenerator = IdGeneratorType.AutoIncrement;
   const schema = new Schema();
   schema.register([docSchema]);
   return { id: 'test-collection', idGenerator, schema };
