+++ packages/framework/store/src/store/collection.ts
@@ -1,27 +1,98 @@
-import { Slot, assertExists } from '@blocksuite/global/utils';
+import {
+  type Logger,
+  NoopLogger,
+  Slot,
+  assertExists,
+} from '@blocksuite/global/utils';
+import {
+  AwarenessEngine,
+  type AwarenessSource,
+  BlobEngine,
+  type BlobSource,
+  DocEngine,
+  type DocSource,
+  MemoryBlobSource,
+  NoopDocSource,
+} from '@blocksuite/sync';
+import { merge } from 'merge';
+import { Awareness } from 'y-protocols/awareness.js';
 import * as Y from 'yjs';
 
 import type { Schema } from '../schema/index.js';
-import type { AwarenessStore } from '../yjs/index.js';
+import type { IdGenerator } from '../utils/id-generator.js';
 import type { BlockSelector, Doc } from './doc/index.js';
+import type { IdGeneratorType } from './id.js';
 
+import {
+  AwarenessStore,
+  BlockSuiteDoc,
+  type RawAwarenessState,
+} from '../yjs/index.js';
 import { DocCollectionAddonType, indexer, test } from './addon/index.js';
 import { BlockCollection, type GetDocOptions } from './doc/block-collection.js';
+import { pickIdGenerator } from './id.js';
 import { DocCollectionMeta, type DocMeta } from './meta.js';
-import { Store, type StoreOptions } from './store.js';
 
-export type DocCollectionOptions = StoreOptions & {
+export type DocCollectionOptions = {
   schema: Schema;
+  id?: string;
+  idGenerator?: IdGeneratorType | IdGenerator;
+  defaultFlags?: Partial<BlockSuiteFlags>;
+  logger?: Logger;
+  docSources?: {
+    main: DocSource;
+    shadows?: DocSource[];
+  };
+  blobSources?: {
+    main: BlobSource;
+    shadows?: BlobSource[];
+  };
+  awarenessSources?: AwarenessSource[];
+  disableSearchIndex?: boolean;
+  disableBacklinkIndex?: boolean;
 };
+
+const FLAGS_PRESET = {
+  enable_synced_doc_block: false,
+  enable_pie_menu: false,
+  enable_database_number_formatting: false,
+  enable_database_statistics: false,
+  enable_database_attachment_note: false,
+  enable_legacy_validation: true,
+  enable_expand_database_block: false,
+  enable_block_query: false,
+  enable_lasso_tool: false,
+  enable_edgeless_text: true,
+  enable_ai_onboarding: false,
+  readonly: {},
+} satisfies BlockSuiteFlags;
+
+export interface StackItem {
+  meta: Map<'cursor-location' | 'selection-state', unknown>;
+}
 
 @indexer
 @test
 export class DocCollection extends DocCollectionAddonType {
   protected readonly _schema: Schema;
 
-  protected _store: Store;
-
   static Y = Y;
+
+  readonly awarenessStore: AwarenessStore;
+
+  readonly awarenessSync: AwarenessEngine;
+
+  readonly blobSync: BlobEngine;
+
+  readonly blockCollections = new Map<string, BlockCollection>();
+
+  readonly doc: BlockSuiteDoc;
+
+  readonly docSync: DocEngine;
+
+  readonly id: string;
+
+  readonly idGenerator: IdGenerator;
 
   meta: DocCollectionMeta;
 
@@ -31,11 +102,47 @@ export class DocCollection extends DocCollectionAddonType {
     docRemoved: new Slot<string>(),
   };
 
-  constructor(options: DocCollectionOptions) {
+  constructor({
+    id,
+    schema,
+    idGenerator,
+    defaultFlags,
+    awarenessSources = [],
+    docSources = {
+      main: new NoopDocSource(),
+    },
+    blobSources = {
+      main: new MemoryBlobSource(),
+    },
+    logger = new NoopLogger(),
+  }: DocCollectionOptions) {
     super();
-    this._schema = options.schema;
+    this._schema = schema;
 
-    this._store = new Store(options);
+    this.id = id || '';
+    this.doc = new BlockSuiteDoc({ guid: id });
+    this.awarenessStore = new AwarenessStore(
+      new Awareness<RawAwarenessState>(this.doc),
+      merge(true, FLAGS_PRESET, defaultFlags)
+    );
+
+    this.awarenessSync = new AwarenessEngine(
+      this.awarenessStore.awareness,
+      awarenessSources
+    );
+    this.docSync = new DocEngine(
+      this.doc,
+      docSources.main,
+      docSources.shadows ?? [],
+      logger
+    );
+    this.blobSync = new BlobEngine(
+      blobSources.main,
+      blobSources.shadows ?? [],
+      logger
+    );
+
+    this.idGenerator = pickIdGenerator(idGenerator, this.doc.clientID);
 
     this.meta = new DocCollectionMeta(this.doc);
     this._bindDocMetaEvents();
@@ -48,9 +155,9 @@ export class DocCollection extends DocCollectionAddonType {
         collection: this,
         doc: this.doc,
         awarenessStore: this.awarenessStore,
-        idGenerator: this._store.idGenerator,
+        idGenerator: this.idGenerator,
       });
-      this._store.addSpace(doc);
+      this.blockCollections.set(doc.id, doc);
       this.slots.docAdded.emit(doc.id);
     });
 
@@ -59,7 +166,7 @@ export class DocCollection extends DocCollectionAddonType {
     this.meta.docMetaRemoved.on(id => {
       const space = this.getBlockCollection(id);
       if (!space) return;
-      this._store.removeSpace(space);
+      this.blockCollections.delete(id);
       space.remove();
       this.slots.docRemoved.emit(id);
     });
@@ -126,7 +233,7 @@ export class DocCollection extends DocCollectionAddonType {
 
     blockCollection.dispose();
     this.meta.removeDocMeta(docId);
-    this._store.removeSpace(blockCollection);
+    this.blockCollections.delete(docId);
   }
 
   /** Update doc meta state. Note that this intentionally does not mutate doc state. */
@@ -158,36 +265,8 @@ export class DocCollection extends DocCollectionAddonType {
     return this.docSync.waitForSynced();
   }
 
-  get awarenessStore(): AwarenessStore {
-    return this._store.awarenessStore;
-  }
-
-  get awarenessSync() {
-    return this.store.awarenessSync;
-  }
-
-  get blobSync() {
-    return this.store.blobSync;
-  }
-
-  get doc() {
-    return this._store.doc;
-  }
-
-  get docSync() {
-    return this.store.docSync;
-  }
-
   get docs() {
-    return this._store.spaces as Map<string, BlockCollection>;
-  }
-
-  get id() {
-    return this._store.id;
-  }
-
-  get idGenerator() {
-    return this._store.idGenerator;
+    return this.blockCollections;
   }
 
   get isEmpty() {
@@ -206,9 +285,5 @@ export class DocCollection extends DocCollectionAddonType {
 
   get schema() {
     return this._schema;
-  }
-
-  get store(): Store {
-    return this._store;
   }
 }
