+++ packages/framework/store/src/store/index.ts
@@ -4,4 +4,3 @@ export type * from './doc/block-collection.js';
 export * from './doc/index.js';
 export * from './id.js';
 export type * from './meta.js';
-export * from './store.js';
