+++ packages/presets/src/__tests__/utils/setup.ts
@@ -3,7 +3,7 @@ import type { BlockCollection } from '@blocksuite/store';
 import { AffineSchemas } from '@blocksuite/blocks/schemas';
 import { assertExists } from '@blocksuite/global/utils';
 import { DocCollection, Text } from '@blocksuite/store';
-import { Generator, Schema } from '@blocksuite/store';
+import { IdGeneratorType, Schema } from '@blocksuite/store';
 
 import { AffineEditorContainer } from '../../index.js';
 
@@ -13,7 +13,7 @@ function createCollectionOptions() {
 
   schema.register(AffineSchemas);
 
-  const idGenerator: Generator = Generator.AutoIncrement; // works only in single user mode
+  const idGenerator: IdGeneratorType = IdGeneratorType.AutoIncrement; // works only in single user mode
 
   return {
     id: room,
